FIX: issues #8: decimal value for worked_hours, and #11: allow duplicates for order and employees. Don't fix #3: order's references with auto-increment.

+++ Intranet-1.0/intranet/tests/model/test_cal_event_data.py
@@ -56,7 +56,7 @@ EMPLOYEE_LIST = [{u'employee_name': u'Bernard',
 
 def make_employee(record):
     return Employee(employee_name=record['employee_name'],
-                    worked_hours=int(record['worked_hours']),
+                    worked_hours=float(record['worked_hours']),
                     entry_date=parse_date(record['entry_date']),
                     exit_date=parse_date(record['exit_date']),
                     photo_path=record['photo_path'])
